A lot of fixes
Removed unnecessary widgets from the inspector for subshader and shadergraph
Added active on Master nodes. Did this to dissallow deleting the active master nodes. Before this you could delete all master nodes.
This also made it so that we show only the active master node in the Main preview.
Also if you are trying to make a subgraph with nodes that are not allowed into subgraphs we prompt the user and let them know that those are not allowed and we can remove them from the selection and continue doing the subgraph.